LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/data-uri-expected.txt


PASS XHR method GET with charset text/plain 
PASS XHR method GET with charset text/plain (base64) 
PASS XHR method GET with charset text/html 
PASS XHR method GET with charset text/html;charset=UTF-8 
PASS XHR method GET with charset image/png 
FAIL XHR method POST with charset text/plain assert_equals: expected 0 but got 200
FAIL XHR method PUT with charset text/plain assert_equals: expected 0 but got 200
FAIL XHR method DELETE with charset text/plain assert_equals: expected 0 but got 200
FAIL XHR method HEAD with charset text/plain assert_equals: expected 0 but got 200
FAIL XHR method UNICORN with charset text/plain assert_equals: expected 0 but got 200
